Added login and registration

+++ base/views.py
@@ -1,7 +1,59 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Recipe, Recipe_Type
 from .forms import RecipeForm
+
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR Password does not exist')
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Error occurred during registration!')
+    return render(request, 'base/login_register.html', {'form': form})
 
 
 def home(request):
@@ -12,6 +64,7 @@ def home(request):
     return render(request, 'base/home.html', context)
 
 
+@login_required(login_url='login')
 def createRecipe(request):
     form = RecipeForm()
     if request.method == 'POST':
@@ -23,9 +76,13 @@ def createRecipe(request):
     return render(request, 'base/recipe_creation.html', context)
 
 
+@login_required(login_url='login')
 def updateRecipe(request, pk):
     recipe = Recipe.objects.get(id=pk)
     form = RecipeForm(instance=recipe)
+
+    if request.user != recipe.user:
+        return HttpResponse('You are not the chef!')
 
     if request.method == 'POST':
         form = RecipeForm(request.POST, instance=recipe)
@@ -36,8 +93,12 @@ def updateRecipe(request, pk):
     return render(request, 'base/recipe_creation.html', context)
 
 
+@login_required(login_url='login')
 def deleteRecipe(request, pk):
     recipe = Recipe.objects.get(id=pk)
+    if request.user != recipe.user:
+        return HttpResponse('You are not the chef!')
+
     if request.method == 'POST':
         recipe.delete()
         return redirect('home')
